Lab9/chat-client-ajfrankman/requirements.txt
click==7.1.2
Flask==2.0.2
paho_mqtt==1.6.1
